requirements.txt
discord.py>=2.3.0
aiohttp>=3.8.0
beautifulsoup4>=4.11.0
python-dotenv>=1.0.0
requests>=2.28.0
undetected-chromedriver==3.5.5
selenium==4.25.0
bs4==0.0.2
